+++ pom.xml
@@ -12,7 +12,7 @@
     <module>dubbo</module>
     <module>spring-boot-dubbo</module>
     <module>spring-cloud-zuul</module>
-    <module>spring-cloud-zuul-retalimit</module>
+    <module>spring-cloud-zuul-ratelimit</module>
     <module>spring-boot-logger</module>
     <module>dynamic-datasource</module>
     <module>spring-boot-encrypt</module>
