+++ spring-cloud-zuul-ratelimit/src/main/java/com/ctrip/framework/apollo/use/cases/spring/cloud/zuul/Application.java
@@ -33,7 +33,7 @@ public class Application {
     app.run(args);
   }
 
-  @GetMapping("/test/index")
+  @GetMapping("/index")
   public String index() {
     return String.format("index %s",  port);
   }
