+++ spring-cloud-zuul-ratelimit/src/main/resources/application.properties
@@ -1,0 +1,18 @@
+app.id=spring-cloud-zuul-ratelimit
+# set apollo meta server address, adjust to actual address if necessary
+apollo.meta=http://localhost:8080
+# will inject 'application' namespace in bootstrap phase https://github.com/ctripcorp/apollo/wiki/Java%E5%AE%A2%E6%88%B7%E7%AB%AF%E4%BD%BF%E7%94%A8%E6%8C%87%E5%8D%97#3213-spring-boot%E9%9B%86%E6%88%90%E6%96%B9%E5%BC%8F%E6%8E%A8%E8%8D%90
+apollo.bootstrap.enabled=true
+# put apollo initialization before logging system initialization
+#apollo.bootstrap.eagerLoad.enabled=true
+# https://github.com/ctripcorp/apollo/wiki/Java%E5%AE%A2%E6%88%B7%E7%AB%AF%E4%BD%BF%E7%94%A8%E6%8C%87%E5%8D%97#3213-spring-boot%E9%9B%86%E6%88%90%E6%96%B9%E5%BC%8F%E6%8E%A8%E8%8D%90
+#apollo.autoUpdateInjectedSpringProperties=true
+server.port=9090
+spring.redis.host=localhost
+spring.redis.port=6379
+spring.redis.database=0
+
+# 如需要可以配置actuator端点，进行手动刷新，非必须
+#management.endpoints.web.exposure.include=refresh,health
+#management.endpoints.web.exposure.include=*
+#management.endpoint.health.show-details=ALWAYS
