--- spring-cloud-zuul-retalimit/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-app.id=spring-cloud-zuul-retalimit
-# set apollo meta server address, adjust to actual address if necessary
-apollo.meta=http://localhost:8080
-# will inject 'application' namespace in bootstrap phase
-apollo.bootstrap.enabled=true
-server.port=9090
-spring.redis.host = localhost
-spring.redis.port = 6379
-spring.redis.database = 0
-
